moved profile component from auth components to components
